requirements.txt
kafka-python
psycopg2-binary
pandas
apache-airflow